# fill_via_chunks: mutate src on BE (small optimisation)