This is an API-breaking change to `rand_core` but not too consequential, and results in a small performance boost on Big Endian:
```
From:
test gen_bytes_chacha12      ... bench:     367,282 ns/iter (+/- 265) = 2788 MB/s
test gen_bytes_chacha20      ... bench:     504,012 ns/iter (+/- 6,271) = 2031 MB/s
test gen_bytes_chacha8       ... bench:     296,971 ns/iter (+/- 1,062) = 3448 MB/s

To:
test gen_bytes_chacha12      ... bench:     334,430 ns/iter (+/- 2,287) = 3061 MB/s
test gen_bytes_chacha20      ... bench:     470,403 ns/iter (+/- 1,755) = 2176 MB/s
test gen_bytes_chacha8       ... bench:     262,490 ns/iter (+/- 1,313) = 3901 MB/s
```

LE is essentially unchanged:
```
test gen_bytes_chacha12      ... bench:     298,710 ns/iter (+/- 2,715) = 3428 MB/s
test gen_bytes_chacha20      ... bench:     436,912 ns/iter (+/- 4,788) = 2343 MB/s
test gen_bytes_chacha8       ... bench:     229,400 ns/iter (+/- 3,040) = 4463 MB/s
```

All numbers are smaller than expected due to #1181.

Since BE is not very popular and it's not that big a bump, we may choose not to merge this.