@dhardy 2021-11-15
Updated with documentation and rebase.

The code is significantly nicer so probably worth merging this (when we accept breaking changes).