@dhardy 2022-02-22
This is a breaking change to `rand_core`, not `rand`. Are we planning a breaking release of that soon? There likely will be before 1.0 (at least @newpavlov has some ideas) but I've no idea what the schedule for that will be (or whether we want another breaking release of `rand_core` for `rand` v0.9).