@vks 2022-02-22
A breaking change to `rand_core` is a breaking change to `rand`, so it would make sense to synchronize them, wouldn't it?