@newpavlov 2022-02-22
One thing which I would like to do before releasing `rand_core v1.0` is changing `BlockRngCore` to:
```rust
pub trait BlockRngCore {
    type Item: SealedTrait;
    const ITEMS: usize;
    fn generate(&mut self) -> [Self::Item; Self::ITEMS];
}
```
It's possible today on Nightly with `generic_const_exprs`, but, unfortunately, it has no clear stabilization timeline.

Such trait still has a certain deficiency around target features, e.g. IIRC with ChaCha we have to use ChaCha block size multiplied by maximum number of blocks processed in parallel (i.e. on non-AVX2 targets code will be less efficient than possible), but we probably can overlook it. In RustCrypto we "solved" it using emulation of rank-2 closures, but such solution is somewhat complex and does not fit well into the `rand` use-case (when you use as a stream cipher you usually consume many blocks at once, while with RNG you often generate `u32`/`u64`s one by one).