@dhardy 2022-02-23
> so it would make sense to synchronize them, wouldn't it?

More to the point, a breaking release of `rand_core` requires a breaking release of `rand`, but not vice versa.

So we can't release `rand_core` v1.0 for a while yet (also dependant on `getrandom`); the question is whether we want to make a breaking release soon or hold off on it for now.

If we didn't make a breaking release of `rand_core`, we wouldn't need to bump the RNG crates either.